refactor(codegen): split let_stmt.rs local-fact helpers under the 2000-line cap
The layout-declaration change pushed let_stmt.rs to 2004 lines, tripping
check_file_size.sh. Moves the POD/native-rep provenance, alias-source,
scalar-class and diagnostic helpers to let_stmt_facts.rs verbatim.
